+++ src/controllers/AdminControllers/AdminPostsController/interfaces/AdminPostsControllerInterface.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rehor\Myblog\controllers\AdminControllers\AdminPostsController\interfaces;
+
+use Rehor\Myblog\controllers\AdminControllers\AdminController\interfaces\AdminControllerInterface;
+
+interface AdminPostsControllerInterface extends AdminControllerInterface
+{
+    public static function showAdminPosts(): void;
+
+    public static function createPosts(): void;
+
+    public static function updatePosts(string $id): void;
+
+    public static function deletePosts(string $id): void;
+}
