+++ src/controllers/BlogController/BlogPostController/BlogPostController.php
@@ -7,7 +7,7 @@ namespace Rehor\Myblog\controllers\BlogController\BlogPostController;
 use Rehor\Myblog\controllers\BlogController\BlogController;
 use Rehor\Myblog\controllers\BlogController\traits\BlogControllerTrait;
 use Rehor\Myblog\controllers\AuthController\AuthController;
-use Rehor\Myblog\controllers\AdminController\AdminController;
+use Rehor\Myblog\controllers\AdminControllers\AdminController\AdminController;
 use Rehor\Myblog\repositories\PostRepository\PostRepository;
 use Rehor\Myblog\repositories\DBConnectorRepositories\DBConnectorFlightRepository\DBConnectorFlightRepository;
 use Rehor\Myblog\repositories\RendererRepository\RendererRepository;
